Remove clang-tidy integration, change to C++20, use ninja
Too lazy to separate this into 3 commits, so this will have to do.

When setting up ninja (since default generator does not compile with all
cores) and changing to C++20 (remove warning for some gtest features), I
observe that the clang-tidy integration broke. The fix is likely to be
complex, so I decide to remove it altogether.