Add presets for using the clang compiler toolchain
I add this because ThreadSanitizer in gcc/g++ have very weird behavior
in my set concurrent tests, which I assume to be a bug (since running
with clang does not have the same issue).

nne other reason for me suspecting a gcc compiler bug is that the stack
trace reported by ThreadSanitizer points to the destructor of
std::threads, which have nothing to do with the custom set data
structure.